hanlder so user cand deleete movie by erquesting id

+++ internal/data/movies.go
@@ -122,8 +122,12 @@ func (m MovieModel) Update(movie *Movie) error {
 }
 
 func (m MovieModel) Delete(id int64) error {
+	if id < 1 {
+		return ErrRecordNotFound
+	}
+
 	query := `DELETE FROM movies
-	WHERE id = $1`
+		WHERE id = $1`
 
 	result, err := m.DB.Exec(query, id)
 	if err != nil {
